FactoryのCertitudeを列挙型から常数の入った構造体へ変更
※列挙でも名前空間に入れたかっただけなのだが、列挙では型変換が必要なので変更

+++ tests/factory.cc
@@ -13,7 +13,7 @@ uint C::r(0);
 
 struct D : C {
 	struct F : tb::Factory<C> {
-		uint Score() override { return (uint)Certitude::passiveMatch; };
+		uint Score() override { return Certitude::passiveMatch; };
 		D* New() override {
 			r = 1;
 			return 0;
@@ -22,7 +22,7 @@ struct D : C {
 };
 struct E : C {
 	struct F : tb::Factory<C> {
-		uint Score() override { return (uint)Certitude::activeMatch; };
+		uint Score() override { return Certitude::activeMatch; };
 		E* New() override {
 			r = 2;
 			return 0;
